用lua脚本实现点赞功能，帖子数据写在redis中，定时写回数据库

+++ src/main/java/com/yupi/friend/model/vo/PostVO.java
@@ -39,9 +39,9 @@ public class PostVO {
 
     private UserVO createUser;
 
-    private Long likeCount;
+    private Integer likeCount;
 
-    private Long commentCount;
+    private Integer commentCount;
 
     private boolean like;
 }
